third_party/blink/web_tests/platform/linux-chrome/external/wpt/wai-aria/role/table-roles-expected.txt

This is a testharness.js-based test.
[FAIL] span role is cell (in div with row role, in div with rowgroup role, in div with table role)
  assert_equals: <span role="cell" data-testname="span role is cell (in div with row role, in div with rowgroup role, in div with table role)" data-expectedrole="cell" class="ex">x</span> expected "cell" but got "gridcell"
[FAIL] orphan span role is cell
  assert_equals: <span role="cell" data-testname="orphan span role is cell" data-expectedrole="cell" class="ex">x</span> expected "cell" but got "gridcell"
Harness: the test ran to completion.
